* create a first SetupDialog (to configure the root node receiver) * create ReceiverConfiguration and ReceiverConfigurator * added a field which describes the receiver status, implemented with OggSource descriptions for the moment * modified the DefaultNodeService to register again the node if the tracker seems to be rebooted * fixed a bug into the hashCode method of CompositeAuditorProvider
git-svn-id: http://svn.tryphon.org/kolaka/freecast/trunk@75 20e14386-5fff-0310-853f-b84fc43df78c

+++ source/org/kolaka/freecast/transport/receiver/StaticBandwidthControler.java
@@ -33,9 +33,9 @@ import org.kolaka.freecast.packet.LogicalPage;
  */
 public class StaticBandwidthControler implements BandwidthControler {
 
-	private int bandwidth;
+	private long bandwidth;
 
-	public StaticBandwidthControler(int bandwidth) {
+	public StaticBandwidthControler(long bandwidth) {
 		Validate.isTrue(bandwidth > 0, "Invalid bandwidth: " + bandwidth);
 		this.bandwidth = bandwidth;
 	}
